feat: adds remote debugging utility
Adds a new utility class, RemoteDebugger, which will configure
containers to start with a debugging server attached.

Note that there are no tests, mostly because any test for this feature
would be contrived I think. At least I couldn't find a simple way to
test something like this in a programmatic way.