build: downgrade project maven setup to JDK 8
Downgrade project build pipeline to Java 8. Since we target Java 8, to
ensure that users on Java 8 can use this library, we need to also run
our complete pipeline over Java 8. Unfortunately targeting the language
level is not enough as that only guarantees newer language features are
not used, but does not guarantee things such as using new standard
library functions which are not available in Java 8.