build: flatten deployed POM
Adds flatten-maven-plugin to the build pipeline. This will generate a
flattened POM, which removes any POM hierarchy, substitutes all
properties, removes non compile dependencies, and more. This solves
issues when it comes to dependency management for consumers of this
project by greatly reducing how much they have to care when it comes to
dependency convergence.